Replace spos by new wrapper around swap_id. Should be improved but seems to work.
This approach is rather ugly in terms of compatibility as in some cases
we want to the value to be treated as one integer and in some cases as
two. A better approach would probably be to just use an integer and
extract the two values using two non-member methods. However, it seems
to work for now.