Make networkit LFR more robust
NetworKit refuses to generate the network when the drawn maximum
degrees/community sizes do not fit. We just re-run the algorithm then
as it should work in a different run. This is not tested properly (i.e.
I did not test this actually fixes the problem), but I did not get any
error anymore with this code.